+++ GeoBot/GeoBot/Healthcheck.cs
@@ -129,97 +129,10 @@ namespace GeoBot
             }
         }
 
-        public async Task<string> GetSpeechToken()
-        {
-            // Check on Speech Endpoint
-            var speechKeyKey = "SpeechAPIKey" + configuration["region"];
-            var speechHostNameKey = "SpeechAPIHostName" + configuration["region"];
-            var speechIsConfigured = !string.IsNullOrEmpty(configuration[speechKeyKey]) && !string.IsNullOrEmpty(configuration[speechHostNameKey]);
-            if (speechIsConfigured)
-            {
-                var hostname = configuration[speechHostNameKey];
-                var speechUrl = $"{hostname}/sts/v1.0/issuetoken";
-                var speechKey = configuration[speechKeyKey];
-
-                using (var client = new HttpClient())
-                {
-                    client.DefaultRequestHeaders.Add("Ocp-Apim-Subscription-Key", speechKey);
-                    UriBuilder uriBuilder = new UriBuilder(speechUrl);
-
-                    var result = await client.PostAsync(uriBuilder.Uri.AbsoluteUri, null);
-                    return await result.Content.ReadAsStringAsync();
-                }
-            }
-            else
-            {
-                // In Case Speech is not configured - skip healthcheck
-                return null;
-            }
-        }
-
-        public string GetSpeechRegion()
-        {
-            var speechKeyKey = "SpeechAPIKey" + configuration["region"];
-            var speechHostNameKey = "SpeechAPIHostName" + configuration["region"];
-            var speechIsConfigured = !string.IsNullOrEmpty(configuration[speechKeyKey]) && !string.IsNullOrEmpty(configuration[speechHostNameKey]);
-            if (speechIsConfigured)
-            {
-                return configuration["region"];
-            }
-            else
-            {
-                // In Case Speech is not configured - skip healthcheck
-                return null;
-            }
-        }
-
-        public async Task<string> GetDirectlineToken()
-        {
-            // Check on Speech Endpoint
-            var directlineIsConfigured = !string.IsNullOrEmpty(configuration["DirectlineKey"]);
-            if (directlineIsConfigured)
-            {
-                var directlineKey = configuration["DirectlineKey"];
-                var directlineUrl = $"https://directline.botframework.com/v3/directline/tokens/generate";
-
-                using (var client = new HttpClient())
-                {
-                    HttpRequestMessage request = new HttpRequestMessage(HttpMethod.Post,directlineUrl);
-                    request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", directlineKey);
-                    var userId = $"dl_{Guid.NewGuid()}";
-
-                    request.Content = new StringContent(
-                        JsonConvert.SerializeObject(
-                            new { User = new { Id = userId } }),
-                            Encoding.UTF8,
-                            "application/json");
-
-                    var result = await client.SendAsync(request);
-                    string token = String.Empty;
-
-                    if (result.IsSuccessStatusCode)
-                    {
-                        var body = await result.Content.ReadAsStringAsync();
-                        token = JsonConvert.DeserializeObject<DirectLineToken>(body).token;
-
-                        return token;
-                    }
-                    else
-                    {
-                        return "There is something wrong when issue the directline token";
-                    }
-                }
-            }
-            else
-            {
-                // In Case Directline is not configured
-                return null;
-            }
-        }
-
         private async Task<HealthcheckResult> CheckSpeech()
         {
-            var speechTokenResult = await GetSpeechToken();
+            Speech speech = new Speech(configuration, null);
+            var speechTokenResult = await speech.GetSpeechToken();
             if (String.IsNullOrEmpty(speechTokenResult))
             {
                 return new HealthcheckResult(true, "200", "Speech not configured");
