+++ GeoBot/GeoBot/Startup.cs
@@ -93,6 +93,12 @@ namespace GeoBot
 
             // Add Healthcheck module
             services.AddSingleton<Healthcheck>();
+
+            // Add Speech module
+            services.AddSingleton<Speech>();
+
+            // Add Directline module
+            services.AddSingleton<Directline>();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
